introducing mime_fetch_headers
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@1527 7b491191-dbf0-0310-aff6-d879d4d69008

+++ imap4.h
@@ -49,7 +49,7 @@ typedef struct {
 	int itemtype;		/* the item to be fetched */
 	int argstart;		/* start index in the arg array */
 	int argcnt;		/* number of args belonging to this bodyfetch */
-	long long octetstart, octetcnt;	/* number of octets to be retrieved */
+	guint64 octetstart, octetcnt;	/* number of octets to be retrieved */
 
 	char partspec[IMAP_MAX_PARTSPEC_LEN];	/* part specifier (i.e. '2.1.3' */
